+++ src/models/api/sources/apple_pay/apple_pay_response_source.ts
@@ -1,4 +1,4 @@
-import PaymentType from '../../../payment_type';
+import { PaymentType } from '../../../payment_type';
 import type { PaymentResponseSource } from '../payment_response_source';
 
 export class ApplePayPaymentResponseSource implements PaymentResponseSource {