+++ src/models/api/sources/credit_card/credit_card_request_source.ts
@@ -1,6 +1,6 @@
 import { getCreditCardNetworkFromNumber } from '../../../../helpers/credit_card_utils';
 import { CreditCardNetwork } from '../../../credit_card_network';
-import PaymentType from '../../../payment_type';
+import { PaymentType } from '../../../payment_type';
 import type { PaymentRequestSource } from '../payment_request_source';
 
 export class CreditCardRequestSource implements PaymentRequestSource {
@@ -20,16 +20,16 @@ export class CreditCardRequestSource implements PaymentRequestSource {
     cvc,
     month,
     year,
-    tokenizeCard,
-    manualPayment,
+    tokenizeCard = false,
+    manualPayment = false,
   }: {
     name: string;
     number: string;
     cvc: string;
     month: string;
     year: string;
-    tokenizeCard: boolean;
-    manualPayment: boolean;
+    tokenizeCard?: boolean;
+    manualPayment?: boolean;
   }) {
     this.network = getCreditCardNetworkFromNumber(number);
     this.name = name;