+++ src/models/component_models/credit_card_props.ts
@@ -1,8 +1,0 @@
-import type { PaymentConfig } from '../payment_config';
-import type { ResultCallback } from '../payment_result';
-
-export interface CreditCardProps {
-  paymentConfig: PaymentConfig;
-  onPaymentResult: ResultCallback;
-  setWebviewVisible: React.Dispatch<React.SetStateAction<boolean>>;
-}