+++ src/models/component_models/moyasar_style.ts
@@ -1,0 +1,17 @@
+import { type TextStyle, type ViewStyle } from 'react-native';
+
+export interface CreditCardMoyasarStyle {
+  container?: ViewStyle;
+  textInputs?: TextStyle;
+  paymentButton?: ViewStyle;
+  paymentButtonText?: TextStyle;
+  errorText?: TextStyle;
+}
+
+export interface ApplePayMoyasarStyle {
+  buttonType?: string; // 'plain' | 'buy' | 'inStore' | 'donate' | 'setUp'
+  buttonStyle?: string; // 'black' | 'white' | 'whiteOutline'
+  height?: string | number;
+  width?: string | number;
+  cornerRadius?: number;
+}