+++ src/models/moyasar_props.ts
@@ -1,7 +1,0 @@
-import type { PaymentConfig } from './payment_config';
-import type { ResultCallback } from './payment_result';
-
-export interface MoyasarProps {
-  paymentConfig: PaymentConfig;
-  onPaymentResult: ResultCallback;
-}