+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.0
+
+- [API] `metadata` parameter type fix.
+- [General] Support customizing the components' style to align with your app's design.
+- [General] Support building your own UI.
+- [General] Enhancements.
+
 ## 0.4.0
 
 - [Credit Card] Support create save only token.